+++ eminus/operators.py
@@ -33,7 +33,7 @@ import numpy as np
 from scipy.fft import fftn, ifftn
 
 from . import config
-from .utils import handle_k, handle_spin, handle_torch
+from .utils import handle_backend, handle_k, handle_spin
 
 
 # Spin handling is trivial for this operator
@@ -110,7 +110,7 @@ def Linv(atoms, W):
     return out
 
 
-@handle_torch
+@handle_backend
 @handle_k(mode="index")
 @handle_spin
 def I(atoms, W, ik=-1):
@@ -162,7 +162,7 @@ def I(atoms, W, ik=-1):
     return Finv
 
 
-@handle_torch
+@handle_backend
 @handle_k(mode="index")
 @handle_spin
 def J(atoms, W, ik=-1, full=True):
@@ -207,7 +207,7 @@ def J(atoms, W, ik=-1, full=True):
     return F
 
 
-@handle_torch
+@handle_backend
 @handle_k(mode="index")
 @handle_spin
 def Idag(atoms, W, ik=-1, full=False):
@@ -233,7 +233,7 @@ def Idag(atoms, W, ik=-1, full=False):
     return F * n
 
 
-@handle_torch
+@handle_backend
 @handle_k(mode="index")
 @handle_spin
 def Jdag(atoms, W, ik=-1):