+++ eminus/utils.py
@@ -216,11 +216,11 @@ def handle_k(func=None, *, mode="gracefully"):
     return decorator
 
 
-def handle_torch(func, *args, **kwargs):
-    """Use a function optimized with Torch if available.
+def handle_backend(func, *args, **kwargs):
+    """Use a function optimized with a different backend if available.
 
     Args:
-        func: Function with a Torch alternative.
+        func: Function with an alternative implementation.
         args: Pass-through arguments.
 
     Keyword Args:
@@ -232,7 +232,10 @@ def handle_torch(func, *args, **kwargs):
 
     @functools.wraps(func)
     def decorator(*args, **kwargs):
-        if config.use_torch:
+        if config.backend == "jax":
+            func_jax = getattr(eminus.extras.jax, func.__name__)
+            return func_jax(*args, **kwargs)
+        if config.backend == "torch":
             func_torch = getattr(eminus.extras.torch, func.__name__)
             return func_torch(*args, **kwargs)
         return func(*args, **kwargs)