+++ tests/test_operators.py
@@ -11,7 +11,7 @@ from numpy.testing import assert_allclose
 
 from eminus import Atoms, config
 
-config.use_torch = False
+config.backend = "scipy"
 
 # Create an Atoms object to build mock wave functions
 atoms = Atoms("Ne", (0, 0, 0), ecut=1).build()