+++ tests/test_utils.py
@@ -13,9 +13,9 @@ from eminus.utils import (
     add_maybe_none,
     atom2charge,
     get_lattice,
+    handle_backend,
     handle_k,
     handle_spin,
-    handle_torch,
     molecule2list,
     pseudo_uniform,
     vector_angle,
@@ -179,18 +179,18 @@ def test_handle_k(mode):
     assert_equal(out, W[0][0])
 
 
-def test_handle_torch():
-    """Test the handle_torch decorator."""
+def test_handle_backend():
+    """Test the handle_backend decorator."""
 
-    @handle_torch
+    @handle_backend
     def mock(x):
         return x
 
-    config.use_torch = False
+    config.backend = "scipy"
     out = mock(np.pi)
     assert_equal(out, np.pi)
-    config.use_torch = True
-    if config.use_torch:
+    config.backend = "torch"
+    if config.backend == "torch":
         with pytest.raises(AttributeError):
             mock(np.pi)